LYOCR/requirements.txt
baidu-aip==2.2.18.0
certifi==2020.12.5
chardet==4.0.0
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
numpy==1.19.5
openpyxl==3.0.6
pandas==1.2.1
Pillow==8.1.0
python-dateutil==2.8.1
pytz==2020.5
pywin32==300
requests==2.25.1
six==1.15.0
tencentcloud-sdk-python==3.0.330
urllib3==1.26.3
